requirements.txt
accelerate
python-dotenv
pyspellchecker
spacy
transformers
